Update brightness-control extension
- docs: update command description and README wording
- refactor: address PR review feedback
