+++ extensions/brightness-control/CHANGELOG.md
@@ -2,11 +2,9 @@
 
 ## [Lunar-based Brightness Control] - {PR_MERGE_DATE}
 
-- Added "Set Brightness" command: set exact brightness level (1-100) on a specific display using Lunar
+- Added "Set Brightness" command: set exact brightness level (1-100) directly from the search bar using Lunar
 - Added "Max Brightness" command: instantly set brightness to 100% on the current display
-- Multi-monitor support with automatic cursor-based display detection
-- Adaptive mode (sync) toggle for non-main displays
-- Guided Lunar installation flow
+- Automatic cursor-based display detection
 
 ## [Initial Version] - 2022-08-03
